backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
pymongo==4.10.1
python-dotenv==1.0.1
openai==1.59.7
PyMuPDF==1.25.1
reportlab==4.2.5
gunicorn==23.0.0
PyMySQL==1.1.1
pyodbc==5.2.0
